+++ NEWS.md
@@ -1,6 +1,7 @@
 # guideR (development version)
 
 * `proportion()` could be applied to atomic vectors (#4)
+* new function `periods_to_long()`
 
 # guideR 0.1.0